refactor(test): migrate local test setup to TypeScript
- The local-test.js script has been rewritten in TypeScript as tests/local-test.ts.
- This migration enhances type safety and maintainability for the local development test environment.
- All related local test assets, including local-template.md and local-test-workflow.yml, have been moved into a dedicated tests/ directory for improved organization.
- Error handling for missing GITHUB_TOKEN and GITHUB_USERNAME environment variables has been improved.
- Adjusted maxPrs configuration in local-test-workflow.yml for the test run.
